+++ RELEASES.md
@@ -6,8 +6,9 @@ Ladybug Test Tool release notes
 Upcoming
 --------
 
-- Make default update behaviour overwrite on XmlStorage
-- Fix unaccepted filename characters on XmlStorage
+- Fix metadata file format for XmlStorage
+- Improve storage id handling of XmlStorage while reading from a filesystem.
+- Fix report discovery for XmlStorage
 - Make security log configurable
 - Upgrade junit from 4.12 to 4.13.1
 - Replace log4j with slf4j